# Much faster validation