--- a/auth-tokens/src/test/java/com/palantir/tokens/auth/UuidStringConverterTests.java
+++ b/auth-tokens/src/test/java/com/palantir/tokens/auth/UuidStringConverterTests.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2018 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.tokens.auth;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.UUID;
+import java.util.stream.IntStream;
+import org.junit.Test;
+
+public class UuidStringConverterTests {
+
+    @Test
+    public void testToString() {
+        IntStream.range(0, 10_000)
+                .forEach(x -> {
+                    UUID uuid = UUID.randomUUID();
+                    assertThat(UuidStringConverter.toString(uuid)).isEqualTo(uuid.toString());
+                });
+    }
+}