To take a String header, convert it into a bearer token, and then turn it into
an unverified JWT takes 5400ns. Additionally the method itself requires the
bearer token, which adds another 3600ns - in total we spend 9us decoding
bearer tokens on each call. This is actually significant; if we want a service
to do 10k RPCs per core, they can expect roughly 9% of all CPU cycles to
be spent decoding bearer tokens.

This PR drops the overhead significantly; it hand codes the validation
of bearer tokens (which drops bearer token parsing to 450ns from 3600ns),
replaces UUID.toString (which is horrible) with a custom implementation
which saves about 450ns on each decode, and lastly replaces the base64
decoder with one from XML which runs 50ns faster (for free).

So, before: 5400ns, after: 1480ns