@j-baker 2018-07-04
Also, while 10us may sound like an eatable amount of error, internally I've noticed a lot of places which casually add 5-10 microseconds, and they can add up. If you factor in metrics + logging, even at 1k requests per second you're spending >20ms per second CPU time doing this work right now, and I'd like to reduce these overheads where they exist.